snap scroll working
