+++ src/components/HomePage/HomeCard.tsx
@@ -1,870 +1,60 @@
+import { Typography, Box, Container } from '@mui/material'
 import React from 'react'
+import { useInView } from 'react-intersection-observer'
 
 export default function HomeCard() {
+  const { ref } = useInView({
+    threshold: 0.75,
+  })
   return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="1055"
-      height="720"
-      fill="none"
+    <div
+      className="snap-child-start"
+      ref={ref}
+      style={{
+        height: '120vh',
+      }}
     >
-      <g filter="url(#a)" shape-rendering="crispEdges">
-        <rect
-          width="549.668"
-          height="453.973"
-          x="61.7591"
-          y="18"
-          fill="url(#b)"
-          rx="40"
-          transform="rotate(3 61.7591 18)"
-        />
-        <rect
-          width="548.668"
-          height="452.973"
-          x="62.2322"
-          y="18.5255"
-          stroke="url(#c)"
-          stroke-opacity=".6"
-          rx="39.5"
-          transform="rotate(3 62.2322 18.5255)"
-        />
-      </g>
-      <path
-        fill="#000"
-        d="m109.623 79.1052-.209 3.9936-12.5095-.6556.2093-3.9936 12.5092.6556Zm-9.7608-21.424-1.3023 24.8487-5.1199-.2683 1.3023-24.8487 5.1199.2683Zm22.4088 26.4336c-1.433-.0751-2.707-.37-3.821-.8848-1.113-.5261-2.047-1.2196-2.8-2.0805-.742-.8603-1.296-1.842-1.66-2.9449-.364-1.1143-.513-2.2859-.449-3.5147l.036-.6826c.073-1.3995.339-2.669.798-3.8087.459-1.1396 1.081-2.1167 1.865-2.9313.796-.8139 1.741-1.4204 2.834-1.8195 1.094-.4104 2.312-.5805 3.655-.5101 1.308.0686 2.457.3456 3.447.831.99.4854 1.806 1.1442 2.446 1.9764.653.8328 1.126 1.816 1.42 2.9495.295 1.1221.407 2.3601.336 3.714l-.107 2.048-14.677-.7692.171-3.2767 9.848.516.019-.3754c.036-.6827-.057-1.2979-.279-1.8458-.211-.5587-.552-1.0101-1.025-1.3543-.472-.3442-1.09-.5363-1.852-.5762-.648-.034-1.213.079-1.695.339-.481.26-.883.6382-1.206 1.1347-.311.4971-.559 1.0888-.743 1.775-.172.6755-.28 1.4229-.323 2.2421l-.036.6826c-.038.7396.028 1.4276.2 2.0641.184.6371.457 1.1933.82 1.6687.374.476.833.8537 1.378 1.1333.556.2801 1.192.4389 1.909.4765.887.0465 1.721-.0809 2.501-.3823.792-.3122 1.486-.8063 2.08-1.4825l2.253 2.7193c-.416.5372-.973 1.0442-1.671 1.521-.687.4774-1.505.8567-2.455 1.1377-.95.2697-2.022.3733-3.217.3106Zm22.327-3.3478.431-8.226c.031-.5916-.039-1.103-.21-1.5341-.171-.4425-.455-.7939-.852-1.0543-.386-.2598-.897-.4064-1.534-.4398-.546-.0286-1.024.0433-1.432.2159-.408.1612-.729.4124-.964.7538-.234.33-.363.7282-.387 1.1947l-4.916-.2576c.042-.785.263-1.5207.664-2.207.401-.6864.962-1.2787 1.684-1.7771.723-.5098 1.57-.8932 2.542-1.1503.983-.2565 2.072-.3535 3.267-.2908 1.433.0751 2.695.3808 3.785.9171 1.09.5363 1.929 1.299 2.516 2.2881.598.9897.86 2.2014.785 3.6349l-.414 7.9018c-.054 1.0126-.034 1.8464.057 2.5015.091.6437.244 1.2108.458 1.7012l-.015.2901-4.966-.2603c-.213-.5131-.362-1.1484-.448-1.9059-.074-.7683-.091-1.5336-.051-2.2959Zm1.02-7.0485-.112 2.7836-2.748-.144c-.648-.034-1.215.0104-1.701.1333-.485.1229-.889.3128-1.21.5697-.321.2456-.565.5409-.731.8858-.155.3456-.244.7289-.266 1.1498-.022.421.055.8072.23 1.1587.176.3401.441.6164.795.8289.355.2011.777.3145 1.266.3401.74.0388 1.39-.0755 1.952-.3427.561-.2671 1.001-.6092 1.32-1.026.33-.4163.511-.8175.542-1.2038l1.188 2.1501c-.206.457-.482.9274-.827 1.4113-.333.4845-.751.9361-1.252 1.3548-.5.4073-1.094.7356-1.78.9849-.685.2378-1.489.3326-2.41.2843-1.172-.0614-2.224-.3504-3.155-.8669-.92-.5274-1.636-1.221-2.15-2.0807-.502-.8706-.725-1.8406-.669-2.9101.051-.9671.278-1.8166.681-2.5484.404-.7319.971-1.3353 1.704-1.8103.744-.4858 1.658-.8316 2.741-1.0372 1.084-.217 2.331-.2885 3.742-.2146l2.85.1494Zm14.919-2.4183-.756 14.4382-4.916-.2576.968-18.4659 4.625.2424.079 4.0429Zm5.781-3.8556-.324 4.5523c-.237-.0467-.526-.0903-.866-.131-.328-.0514-.629-.0843-.902-.0986-.694-.0364-1.301.0231-1.823.1783-.51.1444-.944.3783-1.303.7018-.348.324-.62.7262-.816 1.2065-.185.4809-.306 1.0336-.361 1.6582l-.974-.3591c.063-1.1947.24-2.2863.531-3.275.292-1.0001.685-1.8637 1.18-2.5908.506-.7265 1.105-1.2769 1.798-1.6513.692-.3744 1.471-.539 2.336-.4937.273.0144.551.0517.833.1121.282.0491.513.1125.691.1903Zm8.669 4.5273-.761 14.5235-4.916-.2576.968-18.4659 4.608.2415.101 3.9585Zm-.96 4.6045-1.331-.0698c.071-1.3653.312-2.5849.722-3.6586.41-1.0852.954-1.9979 1.632-2.7382.678-.7516 1.466-1.3093 2.364-1.673.909-.3744 1.91-.5331 3.003-.4758.864.0453 1.648.2119 2.352.4998.704.2878 1.297.7183 1.781 1.2912.494.5736.855 1.3169 1.081 2.2301.237.9137.322 2.0191.254 3.3161l-.617 11.7759-4.95-.2594.618-11.7929c.043-.8192-.037-1.4623-.241-1.9293-.203-.4671-.522-.8032-.957-1.0085-.422-.2161-.952-.3408-1.589-.3742-.66-.0346-1.241.0661-1.744.3022-.492.2366-.909.5798-1.252 1.0296-.331.439-.592.9615-.784 1.5676-.191.606-.305 1.2618-.342 1.9672ZM95.4794 123.425l-3.4879-.183.1394-2.231 3.4655.182c1.0094.053 1.8658-.123 2.5693-.529.711-.405 1.26-1.01 1.647-1.814.3943-.804.6213-1.774.6813-2.911l.046-.886c.046-.882-.014-1.669-.18-2.36-.158-.69-.4201-1.278-.7846-1.761-.357-.484-.8096-.86-1.3578-1.129-.5408-.268-1.1738-.421-1.899-.459l-3.6561-.192.1175-2.243 3.6561.192c1.0842.057 2.0652.292 2.9432.705.8787.406 1.6247.966 2.2397 1.681.622.715 1.084 1.556 1.386 2.524.302.968.422 2.039.36 3.213l-.045.863c-.062 1.174-.293 2.227-.694 3.158-.401.931-.949 1.719-1.642 2.365-.693.639-1.5057 1.117-2.4371 1.436-.9235.311-1.946.438-3.0675.379Zm-1.0732-16.431-.8558 16.33-2.8149-.148.8557-16.329 2.815.147Zm15.8048 15.876 3.864-10.639 2.894.152-5.601 13.73c-.127.293-.29.611-.488.953-.198.342-.451.662-.759.961-.301.307-.67.546-1.106.718-.437.179-.958.253-1.563.222-.24-.013-.47-.048-.692-.104-.215-.049-.419-.101-.611-.156l.097-2.064c.075.011.164.024.268.037.112.013.202.021.269.025.448.023.825-.013 1.13-.11.305-.089.557-.248.756-.477.207-.229.388-.542.544-.939l.998-2.309Zm-1.295-10.909 2.406 9.235.333 2.874-1.898.384L106 111.808l2.916.153Zm13.293 3.294-.5 9.544-2.703-.141.636-12.135 2.546.133.021 2.599Zm-.641 3.003-.874-.057c.053-.859.214-1.642.483-2.347.277-.706.638-1.305 1.084-1.799.453-.494.983-.863 1.588-1.109.605-.253 1.271-.361 1.996-.323.583.031 1.106.141 1.568.33.471.182.864.461 1.182.838.325.377.562.858.712 1.443.149.577.202 1.281.159 2.111l-.411 7.839-2.714-.142.412-7.85c.03-.584-.032-1.048-.186-1.393-.147-.353-.377-.609-.691-.768-.306-.166-.695-.261-1.166-.286-.463-.024-.883.052-1.26.227-.376.175-.701.424-.972.747-.265.324-.476.703-.634 1.137-.157.434-.249.901-.276 1.402Zm18.363 5.056.303-5.787c.023-.434-.036-.812-.176-1.134-.141-.322-.367-.578-.68-.766-.305-.189-.697-.296-1.175-.321-.441-.023-.826.032-1.156.165-.329.132-.59.321-.783.566-.193.245-.298.528-.314.849l-2.692-.141c.025-.478.165-.936.421-1.372.255-.437.612-.819 1.072-1.147.459-.329 1.001-.578 1.625-.747.623-.17 1.316-.235 2.079-.195.912.048 1.712.243 2.398.587.695.344 1.227.836 1.599 1.478.379.635.544 1.416.495 2.343l-.282 5.394c-.029.553-.018 1.053.034 1.498.059.438.155.821.288 1.151l-.009.179-2.771-.145c-.111-.298-.193-.674-.244-1.126-.044-.46-.054-.903-.032-1.329Zm.652-4.925-.065 1.672-1.94-.102c-.501-.026-.945-.001-1.331.077-.386.069-.711.187-.975.353-.263.167-.466.373-.606.621-.14.248-.219.532-.236.854-.017.321.043.62.178.897.136.27.342.487.618.651.284.165.631.258 1.043.28.553.029 1.041-.058 1.464-.261.431-.21.775-.473 1.031-.789.257-.324.401-.635.431-.934l.812 1.246c-.106.302-.276.623-.511.963-.236.34-.537.662-.906.965-.36.296-.788.532-1.285.708-.489.177-1.048.249-1.676.216-.792-.041-1.491-.235-2.095-.582-.603-.354-1.067-.809-1.39-1.366-.323-.564-.466-1.19-.43-1.878.033-.643.183-1.205.448-1.686.273-.488.654-.888 1.142-1.2.497-.311 1.093-.535 1.79-.671.697-.143 1.487-.192 2.369-.145l2.12.111Zm9.119-1.817-.507 9.668-2.703-.142.636-12.135 2.546.134.028 2.475Zm-.603 3.129-.919-.059c.044-.838.189-1.606.436-2.306.246-.699.585-1.3 1.017-1.802.431-.509.953-.891 1.566-1.143.621-.26 1.328-.369 2.121-.328.553.029 1.053.138 1.501.326.455.181.842.453 1.16.814.326.362.565.817.716 1.365.159.548.218 1.203.178 1.966l-.424 8.086-2.703-.142.412-7.85c.031-.591-.035-1.059-.196-1.405-.155-.345-.393-.598-.714-.757-.314-.166-.695-.261-1.144-.285-.508-.026-.947.048-1.316.224-.361.176-.663.426-.905.751-.242.324-.427.704-.555 1.14-.127.436-.204.904-.231 1.405Zm7.563-.323-1.282.214c.039-.733.176-1.419.412-2.059.244-.647.577-1.207.999-1.68.43-.48.947-.846 1.553-1.099s1.29-.36 2.052-.32c.621.032 1.17.147 1.647.345.484.19.889.474 1.215.851.325.377.562.858.711 1.443.15.577.203 1.27.161 2.077l-.411 7.851-2.714-.142.412-7.862c.032-.613-.033-1.093-.195-1.438-.154-.346-.389-.591-.704-.735-.314-.151-.696-.238-1.144-.262-.419-.022-.793.037-1.123.177-.322.133-.598.333-.83.598-.231.258-.411.559-.542.905-.123.346-.195.724-.217 1.136Zm14.516-4.154-.636 12.135-2.714-.143.636-12.134 2.714.142Zm-2.727-3.337c.022-.411.174-.744.458-.999.291-.262.679-.381 1.165-.355.479.025.849.183 1.111.474.263.284.383.631.362 1.042-.021.404-.177.733-.468.988-.291.255-.676.369-1.154.344-.486-.025-.86-.18-1.123-.463-.255-.284-.372-.627-.351-1.031Zm11.258 14.018c.441.023.842-.042 1.203-.196.368-.161.669-.392.902-.695.241-.302.383-.658.427-1.068l2.546.133c-.056.784-.325 1.486-.808 2.106-.482.619-1.103 1.1-1.863 1.443-.76.335-1.589.479-2.486.432-.927-.049-1.726-.248-2.398-.598-.671-.358-1.22-.829-1.646-1.413-.427-.585-.737-1.25-.93-1.995-.186-.744-.257-1.535-.213-2.373l.02-.392c.044-.837.197-1.617.46-2.338.271-.728.649-1.36 1.134-1.897.486-.537 1.081-.944 1.785-1.222.705-.286 1.517-.404 2.436-.356.972.051 1.814.29 2.527.717.713.42 1.26.985 1.643 1.694.39.703.569 1.507.537 2.413l-2.546-.134c.001-.45-.09-.859-.273-1.229-.175-.376-.436-.682-.784-.918-.34-.235-.757-.366-1.25-.392-.546-.028-1.004.06-1.375.266-.37.198-.666.482-.888.853-.221.363-.389.778-.504 1.244-.106.46-.173.94-.199 1.441l-.021.392c-.026.501-.01.989.047 1.465.058.475.178.905.36 1.289.19.377.459.688.806.931.348.235.798.368 1.351.397Zm21.749-3.033-4.25-.222.117-2.232 4.25.223c.74.038 1.345-.05 1.813-.265.469-.216.819-.527 1.05-.935.239-.415.373-.895.402-1.441.027-.516-.056-1.004-.25-1.464-.185-.467-.499-.85-.94-1.151-.442-.301-1.033-.47-1.773-.509l-3.387-.178-.738 14.087-2.815-.148.856-16.329 6.202.325c1.263.066 2.324.347 3.183.842.867.487 1.511 1.136 1.934 1.945.423.802.608 1.704.555 2.706-.055 1.054-.334 1.947-.837 2.677-.504.731-1.211 1.275-2.121 1.632-.904.358-1.987.504-3.251.437Zm12.256-3.102-.514 9.824-2.703-.141.636-12.135 2.579.135.002 2.317Zm3.838-2.194-.154 2.511c-.163-.039-.341-.071-.535-.096-.186-.025-.373-.042-.56-.052-.463-.024-.874.022-1.233.138-.358.109-.663.28-.915.514-.244.227-.439.51-.584.847-.145.337-.24.718-.285 1.143l-.619.013c.04-.763.152-1.465.335-2.108.184-.643.438-1.203.763-1.681.332-.477.734-.842 1.205-1.095.478-.252 1.019-.363 1.625-.331.165.009.339.033.525.073.193.04.337.081.432.124Zm1.705 6.387.014-.258c.046-.875.215-1.68.509-2.414.293-.742.694-1.377 1.201-1.905.516-.535 1.129-.938 1.841-1.208.719-.277 1.52-.393 2.402-.347.89.047 1.674.246 2.353.596.687.343 1.258.808 1.715 1.394.457.579.793 1.253 1.007 2.021.215.761.3 1.579.254 2.454l-.013.258c-.046.875-.216 1.679-.509 2.414-.294.734-.698 1.369-1.213 1.904-.515.528-1.129.931-1.841 1.208-.711.27-1.508.382-2.39.336-.89-.047-1.678-.242-2.365-.585-.679-.351-1.247-.815-1.704-1.394-.456-.586-.792-1.26-1.007-2.021-.215-.761-.3-1.579-.254-2.453Zm2.717-.117-.014.258c-.029.546 0 1.065.087 1.557s.24.927.46 1.306c.22.379.512.683.875.912.362.229.802.357 1.318.384.501.026.94-.056 1.317-.246.385-.19.707-.462.965-.816.258-.353.457-.77.594-1.251.145-.479.232-.992.261-1.538l.013-.258c.028-.538-.005-1.05-.099-1.534-.087-.492-.244-.932-.471-1.318-.22-.387-.511-.694-.874-.923-.355-.236-.787-.368-1.295-.394-.508-.027-.955.058-1.34.256-.378.19-.696.465-.955.827-.258.361-.457.782-.594 1.262-.138.472-.22.978-.248 1.516Zm20.076-4.908 2.457.129-.619 11.798c-.057 1.092-.337 2.007-.841 2.745-.503.738-1.18 1.284-2.031 1.637-.851.36-1.823.511-2.914.454-.464-.024-.976-.118-1.537-.283-.554-.164-1.088-.409-1.604-.736-.508-.319-.925-.727-1.251-1.224l1.351-1.526c.406.539.865.941 1.376 1.208.511.267 1.054.415 1.63.445.62.033 1.153-.056 1.599-.265.453-.201.811-.516 1.073-.944.262-.429.41-.961.443-1.596l.478-9.107.39-2.735Zm-8.568 5.77.013-.235c.048-.92.204-1.752.468-2.495.264-.751.62-1.388 1.067-1.912.448-.531.978-.927 1.592-1.187.614-.268 1.298-.382 2.053-.343.785.041 1.447.219 1.985.532.546.313.992.745 1.338 1.296.347.543.605 1.186.776 1.93.179.736.298 1.552.356 2.447l-.04.752c-.142.862-.348 1.646-.618 2.351-.269.706-.608 1.311-1.017 1.814-.409.503-.901.882-1.477 1.137-.568.255-1.229.363-1.985.323-.74-.038-1.401-.227-1.983-.565-.575-.337-1.061-.794-1.458-1.369-.389-.575-.677-1.243-.862-2.002-.185-.767-.254-1.592-.208-2.474Zm2.716-.094-.013.236c-.029.553-.004 1.072.076 1.556.087.484.233.915.438 1.293.213.371.49.671.83.899.349.22.766.343 1.252.369.635.033 1.162-.074 1.58-.322.425-.248.758-.594.999-1.039.249-.452.433-.959.552-1.523l.107-2.03c-.022-.443-.094-.859-.216-1.248-.115-.388-.281-.731-.497-1.027-.217-.304-.497-.543-.84-.719-.343-.183-.753-.287-1.232-.312-.486-.025-.914.057-1.284.248-.369.19-.68.463-.931.817-.243.355-.434.772-.572 1.252-.137.48-.22.997-.249 1.55Zm14.854-2.887-.515 9.824-2.703-.142.636-12.134 2.579.135.003 2.317Zm3.837-2.195-.154 2.511c-.163-.038-.341-.07-.535-.095-.186-.025-.372-.042-.559-.052-.464-.024-.875.022-1.233.138-.358.108-.664.28-.916.514-.244.227-.439.509-.584.847-.145.337-.24.718-.285 1.143l-.619.013c.04-.763.152-1.466.335-2.108.184-.643.438-1.203.763-1.681.333-.477.734-.842 1.205-1.095.478-.253 1.02-.363 1.625-.331.165.009.34.033.525.072.193.041.337.082.432.124Zm8.767 10.266.303-5.787c.023-.433-.036-.811-.177-1.134-.14-.322-.367-.577-.679-.766-.305-.189-.697-.295-1.175-.321-.441-.023-.827.032-1.156.165-.329.132-.59.321-.783.566-.193.245-.298.528-.315.849l-2.691-.141c.025-.478.165-.936.42-1.372.256-.436.613-.819 1.073-1.147.459-.328 1.001-.577 1.624-.747.624-.17 1.317-.235 2.08-.195.912.048 1.711.244 2.398.587.694.344 1.227.836 1.598 1.478.379.635.545 1.416.496 2.343l-.283 5.394c-.029.554-.017 1.053.034 1.498.06.438.156.821.289 1.151l-.01.179-2.77-.145c-.112-.298-.193-.674-.244-1.126-.044-.46-.054-.903-.032-1.329Zm.652-4.925-.065 1.672-1.941-.102c-.501-.026-.944 0-1.331.077-.386.07-.711.188-.974.354-.264.166-.466.373-.606.62-.141.248-.219.532-.236.854-.017.321.042.62.178.897.136.27.342.487.618.651.284.165.631.258 1.042.28.554.029 1.042-.058 1.465-.261.431-.21.774-.473 1.031-.789.257-.324.4-.635.431-.933l.812 1.245c-.106.302-.276.623-.512.963-.235.34-.537.662-.905.965-.36.296-.789.532-1.285.709-.489.176-1.048.248-1.676.215-.793-.041-1.491-.235-2.095-.582-.604-.354-1.067-.809-1.39-1.366-.323-.564-.467-1.19-.43-1.878.033-.643.183-1.205.448-1.686.273-.488.654-.888 1.142-1.2.496-.311 1.093-.535 1.79-.67.697-.144 1.487-.192 2.369-.146l2.12.111Zm9.119-1.816-.507 9.667-2.703-.142.636-12.134 2.546.133.028 2.476Zm-.603 3.128-.919-.059c.044-.838.189-1.606.436-2.305.246-.7.585-1.3 1.016-1.803.432-.509.954-.891 1.567-1.143.621-.26 1.328-.369 2.12-.328.554.029 1.054.138 1.501.326.456.182.843.453 1.161.815.326.362.565.816.716 1.364.158.548.218 1.204.178 1.966l-.424 8.086-2.703-.141.412-7.851c.031-.591-.035-1.059-.197-1.405-.154-.345-.392-.598-.713-.757-.314-.166-.695-.261-1.144-.285-.508-.026-.947.048-1.316.224-.362.176-.663.426-.905.751-.242.324-.427.705-.555 1.14-.128.436-.205.904-.231 1.405Zm7.563-.323-1.282.214c.038-.733.176-1.419.411-2.059.244-.647.577-1.207.999-1.68.43-.48.948-.846 1.554-1.099.605-.253 1.289-.36 2.052-.32.621.033 1.169.148 1.646.345.485.19.89.474 1.215.851.325.377.563.858.712 1.443.15.578.203 1.27.161 2.078l-.412 7.85-2.714-.142.412-7.862c.033-.613-.032-1.092-.194-1.438-.154-.346-.389-.59-.704-.734-.314-.152-.696-.239-1.144-.263-.419-.022-.793.037-1.123.178-.322.133-.599.332-.83.597-.231.258-.412.559-.542.905-.123.346-.196.725-.217 1.136Zm14.196-1.697-.507 9.668-2.703-.142.636-12.135 2.546.134.028 2.475Zm-.603 3.129-.919-.06c.044-.837.189-1.605.436-2.305.246-.699.585-1.3 1.016-1.802.432-.51.954-.891 1.567-1.144.621-.259 1.328-.369 2.12-.327.554.029 1.054.138 1.502.326.455.181.842.453 1.16.814.326.362.565.817.716 1.365.159.548.218 1.203.178 1.966l-.424 8.086-2.703-.142.412-7.85c.031-.591-.035-1.059-.196-1.405-.155-.345-.393-.598-.714-.757-.314-.166-.695-.261-1.144-.285-.508-.027-.947.048-1.316.223-.362.176-.663.427-.905.751-.242.325-.427.705-.555 1.141-.128.436-.205.904-.231 1.405Zm7.563-.324-1.282.214c.038-.732.176-1.419.412-2.058.243-.647.576-1.207.999-1.68.43-.48.947-.846 1.553-1.099.605-.254 1.29-.36 2.052-.32.621.032 1.169.147 1.646.345.485.19.89.474 1.215.851.326.377.563.858.712 1.443.15.577.203 1.27.161 2.077l-.411 7.851-2.714-.142.412-7.862c.032-.613-.033-1.093-.195-1.439-.154-.345-.389-.59-.704-.734-.314-.151-.696-.239-1.144-.262-.419-.022-.793.037-1.123.177-.322.133-.598.333-.83.598-.231.258-.411.559-.542.905-.123.346-.195.724-.217 1.135Zm14.516-4.153-.636 12.134-2.714-.142.636-12.134 2.714.142Zm-2.727-3.337c.022-.411.174-.744.457-.999.292-.263.68-.381 1.166-.355.479.025.849.183 1.111.474.263.284.383.631.362 1.042-.022.404-.178.733-.468.988-.291.255-.676.369-1.154.344-.486-.025-.86-.18-1.123-.463-.255-.284-.372-.627-.351-1.031Zm9.381 6.283-.501 9.544-2.702-.141.636-12.135 2.545.133.022 2.599Zm-.641 3.003-.875-.057c.053-.859.214-1.642.483-2.348.277-.705.639-1.304 1.084-1.798.453-.494.983-.864 1.588-1.109.605-.253 1.271-.361 1.996-.323.583.031 1.106.141 1.568.33.471.182.865.461 1.182.838.325.377.563.858.712 1.443.15.577.203 1.281.159 2.111l-.411 7.839-2.714-.142.412-7.851c.03-.583-.032-1.047-.186-1.392-.146-.353-.377-.609-.691-.768-.306-.166-.695-.261-1.166-.286-.463-.024-.883.051-1.26.227-.376.175-.7.424-.972.747-.265.324-.476.702-.633 1.137-.158.434-.25.901-.276 1.402Zm19.802-4.596 2.456.128-.619 11.799c-.057 1.091-.337 2.006-.841 2.745-.503.738-1.18 1.283-2.031 1.636-.851.36-1.822.512-2.914.455-.464-.025-.976-.119-1.537-.283-.554-.164-1.088-.41-1.603-.737-.509-.319-.926-.727-1.252-1.223l1.351-1.527c.406.539.865.942 1.376 1.208.511.267 1.054.416 1.63.446.62.032 1.153-.056 1.599-.265.453-.201.811-.516 1.073-.945.262-.428.41-.96.444-1.596l.477-9.107.391-2.734Zm-8.568 5.77.012-.236c.048-.92.204-1.751.468-2.494.264-.751.62-1.389 1.067-1.913.448-.531.979-.927 1.592-1.187.614-.268 1.298-.382 2.053-.342.785.041 1.447.218 1.986.531.545.314.991.746 1.337 1.296.347.543.605 1.186.776 1.93.179.737.298 1.553.356 2.448l-.04.751c-.142.862-.348 1.646-.618 2.352-.269.706-.608 1.31-1.017 1.814-.409.503-.901.882-1.476 1.137-.569.255-1.23.363-1.985.323-.741-.039-1.402-.227-1.984-.565-.574-.338-1.06-.794-1.458-1.37-.389-.575-.677-1.242-.862-2.002-.184-.767-.254-1.591-.207-2.473Zm2.715-.094-.013.235c-.029.554-.003 1.072.076 1.556.087.485.233.916.439 1.294.212.371.489.67.83.898.348.221.765.344 1.251.369.636.034 1.162-.074 1.58-.322.425-.247.758-.593.999-1.038.249-.452.433-.96.552-1.523l.107-2.03c-.022-.444-.094-.86-.216-1.248-.115-.389-.28-.731-.497-1.027-.217-.304-.497-.544-.84-.719-.343-.183-.753-.287-1.232-.312-.486-.026-.914.057-1.283.247-.37.191-.681.463-.932.818-.243.354-.434.772-.571 1.252-.138.48-.221.996-.25 1.55Z"
-      />
-      <rect
-        width="200"
-        height="13"
-        x="87.8719"
-        y="146.058"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 87.8719 146.058)"
-      />
-      <rect
-        width="300"
-        height="13"
-        x="86.3018"
-        y="176.016"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 86.3018 176.016)"
-      />
-      <rect
-        width="230"
-        height="13"
-        x="84.7318"
-        y="205.975"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 84.7318 205.975)"
-      />
-      <rect
-        width="60"
-        height="13"
-        x="83.0046"
-        y="238.93"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 83.0046 238.93)"
-      />
-      <rect
-        width="140"
-        height="13"
-        x="159.899"
-        y="242.96"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 159.899 242.96)"
-      />
-      <rect
-        width="150"
-        height="13"
-        x="314.687"
-        y="251.072"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 314.687 251.072)"
-      />
-      <rect
-        width="87"
-        height="13"
-        x="479.461"
-        y="259.708"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 479.461 259.708)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="73.5319"
-        y="419.682"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 73.5319 419.682)"
-      />
-      <rect
-        width="60"
-        height="13"
-        x="220.33"
-        y="427.375"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 220.33 427.375)"
-      />
-      <rect
-        width="120"
-        height="13"
-        x="297.225"
-        y="431.405"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 297.225 431.405)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="434.037"
-        y="438.575"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 434.037 438.575)"
-      />
-      <rect
-        width="100"
-        height="13"
-        x="305.573"
-        y="157.467"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 305.573 157.467)"
-      />
-      <rect
-        width="100"
-        height="13"
-        x="468.205"
-        y="226.072"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 468.205 226.072)"
-      />
-      <rect
-        width="150"
-        height="13"
-        x="419.417"
-        y="163.433"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 419.417 163.433)"
-      />
-      <rect
-        width="176"
-        height="13"
-        x="402.867"
-        y="192.607"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 402.867 192.607)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="326.4"
-        y="218.641"
-        fill="#788C92"
-        rx="5"
-        transform="rotate(3 326.4 218.641)"
-      />
-      <rect
-        width="484"
-        height="125"
-        x="79.7665"
-        y="276.224"
-        fill="#5B5C5C"
-        rx="10"
-        transform="rotate(3 79.7665 276.224)"
-      />
-      <rect
-        width="140"
-        height="10"
-        x="88.7103"
-        y="315.746"
-        fill="#A4B3B8"
-        rx="5"
-        transform="rotate(3 88.7103 315.746)"
-      />
-      <rect
-        width="100"
-        height="10"
-        x="97.5976"
-        y="337.241"
-        fill="#E03B0A"
-        rx="5"
-        transform="rotate(3 97.5976 337.241)"
-      />
-      <rect
-        width="80"
-        height="10"
-        x="206.448"
-        y="342.945"
-        fill="#E03B0A"
-        rx="5"
-        transform="rotate(3 206.448 342.945)"
-      />
-      <rect
-        width="60"
-        height="10"
-        x="96.5508"
-        y="357.213"
-        fill="#10D146"
-        rx="5"
-        transform="rotate(3 96.5508 357.213)"
-      />
-      <rect
-        width="140"
-        height="10"
-        x="166.455"
-        y="360.877"
-        fill="#10D146"
-        rx="5"
-        transform="rotate(3 166.455 360.877)"
-      />
-      <rect
-        width="165"
-        height="10"
-        x="95.5042"
-        y="377.186"
-        fill="#E03B0A"
-        rx="5"
-        transform="rotate(3 95.5042 377.186)"
-      />
-      <rect
-        width="60"
-        height="10"
-        x="238.505"
-        y="323.596"
-        fill="#A4B3B8"
-        rx="5"
-        transform="rotate(3 238.505 323.596)"
-      />
-      <rect
-        width="120"
-        height="10"
-        x="308.409"
-        y="327.26"
-        fill="#A4B3B8"
-        rx="5"
-        transform="rotate(3 308.409 327.26)"
-      />
-      <rect
-        width="40"
-        height="10"
-        x="435.235"
-        y="333.907"
-        fill="#A4B3B8"
-        rx="5"
-        transform="rotate(3 435.235 333.907)"
-      />
-      <path
-        fill="#A7D0DD"
-        d="m97.4535 301.289 3.5795 3.161-1.6624 1.311-3.513-3.128 1.5959-1.344Zm4.0575-5.4-.041.78c-.056 1.073-.246 2.027-.57 2.865-.324.837-.762 1.541-1.3138 2.111-.5514.571-1.1963.997-1.9347 1.278-.738.274-1.5459.388-2.4236.342-.8712-.045-1.6594-.243-2.3648-.593-.6984-.356-1.2953-.847-1.7906-1.473-.4952-.626-.8668-1.372-1.1148-2.239-.2414-.867-.334-1.836-.2778-2.909l.0409-.78c.0562-1.073.2495-2.024.5799-2.855.3307-.836.7718-1.54 1.3232-2.111.5583-.577 1.2032-1.003 1.9348-1.277.7383-.281 1.5431-.399 2.4143-.353.8777.046 1.669.247 2.374.603.7118.351 1.3119.842 1.8 1.474.489.625.851 1.371 1.086 2.237.242.86.334 1.827.278 2.9Zm-2.4982.651.0419-.8c.0416-.793.0002-1.496-.1242-2.109-.1175-.619-.315-1.141-.5926-1.566-.2708-.431-.6153-.762-1.0335-.993-.4113-.237-.89-.369-1.4362-.398-.5331-.028-1.0166.054-1.4504.247-.4335.186-.8107.479-1.1316.879-.3205.394-.5748.893-.7629 1.496-.1877.596-.3023 1.291-.3439 2.084l-.0419.8c-.0416.793-.0004 1.499.1236 2.118.124.62.3245 1.149.6014 1.587.2838.432.6313.766 1.0427 1.003.4182.231.8939.36 1.427.388.5461.028 1.0359-.05 1.4694-.236.4339-.193.808-.489 1.1223-.89.3147-.407.5629-.912.7445-1.516.188-.603.3028-1.301.3444-2.094Zm11.8362 4.855.423-8.065 2.36.123-.553 10.552-2.224-.116-.006-2.494Zm.447-2.177.791.022c-.037.708-.15 1.358-.337 1.948-.187.583-.455 1.087-.803 1.512-.348.419-.782.739-1.302.959-.52.214-1.134.302-1.843.265-.513-.027-.981-.126-1.402-.298-.421-.172-.78-.422-1.075-.751-.29-.328-.506-.747-.648-1.256-.143-.51-.196-1.112-.16-1.808l.357-6.817 2.351.123-.359 6.837c-.02.383.009.708.086.972.078.259.191.47.339.634.148.164.324.285.528.36.205.076.424.12.658.133.67.035 1.204-.068 1.601-.308.404-.246.699-.589.885-1.029.193-.44.303-.939.333-1.498Zm9.832 5.405c-.78-.041-1.479-.205-2.096-.492-.611-.292-1.125-.681-1.543-1.166-.411-.484-.718-1.045-.919-1.681-.201-.637-.283-1.316-.245-2.037l.02-.39c.043-.826.203-1.567.478-2.225.276-.657.641-1.211 1.095-1.663.454-.459.981-.8 1.579-1.023.598-.222 1.239-.316 1.922-.28.754.039 1.407.201 1.959.484.553.283 1.002.665 1.348 1.146.354.475.605 1.033.754 1.673.155.64.213 1.341.173 2.102l-.052 1.004-8.153-.427.088-1.687 5.832.305.01-.185c.009-.423-.055-.824-.192-1.203-.13-.378-.345-.69-.645-.934-.3-.243-.71-.379-1.23-.406-.39-.021-.743.046-1.057.199-.308.146-.57.371-.788.672-.218.302-.394.671-.528 1.107-.127.43-.204.918-.233 1.464l-.02.39c-.025.462.015.894.117 1.297.11.397.281.748.512 1.054.232.305.519.552.861.739.343.181.738.284 1.187.307.566.03 1.076-.058 1.53-.262.454-.204.854-.506 1.201-.905l1.176 1.264c-.245.32-.558.623-.938.909-.38.28-.838.501-1.374.662-.53.161-1.14.224-1.829.188Zm13.134-2.373c.013-.234-.035-.448-.142-.643-.107-.201-.322-.388-.645-.562-.317-.173-.791-.341-1.421-.505-.553-.152-1.059-.326-1.518-.519-.452-.2-.838-.432-1.157-.696-.318-.265-.56-.571-.724-.919-.164-.347-.235-.742-.212-1.184.023-.429.138-.831.347-1.204.209-.374.496-.698.863-.972.366-.274.804-.483 1.313-.626.516-.142 1.086-.197 1.711-.164.884.046 1.633.235 2.248.567.622.326 1.089.752 1.4 1.277.312.518.452 1.092.418 1.723l-2.35-.123c.015-.28-.043-.543-.173-.791-.124-.255-.322-.464-.593-.628-.272-.171-.619-.267-1.041-.289-.404-.021-.742.026-1.015.142-.267.11-.47.262-.611.457-.134.195-.207.413-.22.653-.009.176.015.337.073.483.064.141.174.274.331.399.156.119.372.235.647.347.28.113.633.226 1.057.339.798.211 1.479.465 2.043.762.572.291 1.003.653 1.293 1.085.291.426.419.954.386 1.585-.024.468-.147.892-.369 1.271-.215.374-.519.694-.911.96-.392.26-.855.454-1.39.582-.528.129-1.118.176-1.768.142-.955-.05-1.756-.261-2.401-.634-.645-.379-1.126-.845-1.443-1.396-.31-.557-.45-1.125-.419-1.704l2.272.119c.003.437.105.791.306 1.063.208.265.471.464.79.599.326.127.664.201 1.016.219.422.022.779-.015 1.071-.11.292-.102.518-.247.678-.434.16-.194.247-.417.26-.671Zm10.599-7.16-.09 1.716-5.949-.311.09-1.717 5.949.312Zm-4.098-2.806 2.351.123-.536 10.22c-.017.325.016.578.097.758.089.174.216.295.382.363.166.067.363.107.591.119.162.008.319.007.469-.005.151-.012.272-.025.364-.04l-.084 1.795c-.198.048-.429.089-.691.12-.256.033-.55.04-.881.023-.54-.029-1.013-.148-1.419-.358-.406-.217-.714-.549-.926-.997-.211-.448-.298-1.033-.26-1.755l.543-10.366Zm9.426 3.085-.553 10.552-2.36-.124.553-10.552 2.36.124Zm-2.371-2.901c.019-.358.151-.648.398-.869.253-.228.591-.331 1.013-.309.417.022.739.159.967.412.228.247.333.549.314.907-.019.351-.154.637-.407.859-.253.221-.587.321-1.003.299-.423-.022-.748-.157-.977-.403-.221-.247-.323-.545-.305-.896Zm5.213 8.458.012-.224c.04-.761.187-1.461.442-2.099.256-.645.604-1.198 1.046-1.657.448-.465.981-.815 1.6-1.05.625-.241 1.321-.342 2.089-.302.773.041 1.455.214 2.046.518.597.299 1.094.703 1.491 1.213.397.503.689 1.089.876 1.757.187.662.26 1.373.221 2.134l-.012.224c-.04.761-.187 1.46-.443 2.099-.255.638-.606 1.19-1.055 1.656-.447.459-.981.809-1.6 1.05-.618.235-1.311.333-2.078.292-.774-.04-1.459-.21-2.057-.508-.59-.305-1.084-.709-1.481-1.212-.397-.51-.689-1.096-.876-1.757-.187-.662-.261-1.373-.221-2.134Zm2.362-.101-.012.224c-.024.475.001.926.076 1.354.076.427.209.806.4 1.135.192.33.445.594.761.793.315.199.697.311 1.146.334.435.023.817-.048 1.145-.214.335-.165.615-.401.839-.709.225-.307.397-.67.517-1.087.126-.418.202-.864.227-1.338l.011-.224c.025-.469-.004-.913-.086-1.335-.076-.428-.212-.81-.41-1.146-.191-.336-.444-.603-.759-.803-.309-.205-.684-.319-1.127-.342-.442-.023-.83.051-1.165.222-.328.166-.605.405-.83.72-.225.314-.397.68-.517 1.097-.119.411-.191.851-.216 1.319Zm12.913-2.247-.435 8.299-2.35-.123.553-10.552 2.213.116.019 2.26Zm-.558 2.611-.76-.05c.046-.747.186-1.428.42-2.041.241-.614.555-1.135.943-1.564.394-.43.854-.751 1.38-.965.527-.22 1.105-.313 1.736-.28.507.026.962.122 1.364.286.409.159.751.402 1.028.729.282.328.489.746.618 1.255.131.502.177 1.114.139 1.836l-.357 6.816-2.36-.123.357-6.827c.027-.507-.027-.911-.161-1.211-.127-.307-.328-.529-.601-.667-.266-.145-.604-.227-1.014-.249-.403-.021-.768.045-1.095.197-.328.152-.61.369-.846.65-.23.281-.413.611-.551.988-.137.378-.217.784-.24 1.22Z"
-      />
-      <g filter="url(#d)" shape-rendering="crispEdges">
-        <rect
-          width="549.668"
-          height="453.973"
-          x="467.759"
-          y="238"
-          fill="url(#e)"
-          rx="40"
-          transform="rotate(3 467.759 238)"
-        />
-        <rect
-          width="548.668"
-          height="452.973"
-          x="468.232"
-          y="238.525"
-          stroke="url(#f)"
-          stroke-opacity=".6"
-          rx="39.5"
-          transform="rotate(3 468.232 238.525)"
-        />
-      </g>
-      <path
-        fill="#fff"
-        d="M512.717 296.437c.023-.443-.024-.845-.142-1.205-.106-.371-.328-.713-.665-1.028-.337-.325-.816-.653-1.437-.982-.622-.329-1.431-.68-2.427-1.051-1.107-.423-2.155-.889-3.144-1.397-.977-.508-1.837-1.083-2.579-1.727-.73-.654-1.296-1.397-1.697-2.228-.39-.83-.557-1.78-.501-2.85.054-1.035.331-1.962.83-2.78.5-.83 1.175-1.524 2.026-2.084.851-.572 1.843-.993 2.976-1.265 1.143-.271 2.386-.371 3.729-.301 1.832.096 3.413.51 4.744 1.242 1.33.731 2.341 1.691 3.032 2.879.702 1.19 1.014 2.524.937 4.003l-5.086-.267c.038-.728-.082-1.373-.361-1.935-.266-.573-.698-1.035-1.296-1.386-.587-.35-1.34-.549-2.262-.598-.887-.046-1.634.051-2.24.294-.605.23-1.068.565-1.388 1.005-.319.428-.493.921-.522 1.479-.022.421.06.807.247 1.159.198.353.5.688.906 1.006.405.318.908.624 1.509.918.601.293 1.298.587 2.093.879 1.333.481 2.497 1.004 3.494 1.57 1.009.566 1.844 1.186 2.504 1.859.661.674 1.146 1.424 1.457 2.25.31.826.439 1.751.385 2.775-.057 1.081-.318 2.037-.784 2.869-.465.831-1.112 1.521-1.939 2.071-.828.55-1.801.95-2.921 1.199-1.12.25-2.362.339-3.728.267-1.228-.064-2.432-.287-3.61-.668-1.177-.393-2.238-.945-3.182-1.656-.932-.711-1.662-1.593-2.188-2.647-.527-1.055-.754-2.282-.68-3.681l5.137.269c-.041.774.044 1.434.255 1.981.211.548.524.998.939 1.35.427.353.938.62 1.533.799.606.18 1.274.29 2.002.328.887.046 1.622-.041 2.204-.261.594-.22 1.038-.545 1.335-.974.307-.429.476-.922.505-1.48Zm9.39-2.058.019-.358c.071-1.354.33-2.59.776-3.707.447-1.129 1.063-2.095 1.846-2.898.784-.803 1.728-1.41 2.834-1.82 1.106-.421 2.358-.595 3.758-.521 1.399.073 2.633.377 3.7.912 1.068.524 1.949 1.226 2.645 2.107.706.881 1.223 1.906 1.55 3.076.327 1.158.455 2.414.384 3.768l-.019.358c-.07 1.343-.329 2.579-.776 3.707-.446 1.118-1.067 2.084-1.863 2.898-.784.803-1.728 1.409-2.833 1.819-1.106.41-2.358.578-3.758.505-1.399-.073-2.633-.372-3.701-.896-1.056-.523-1.938-1.225-2.644-2.106-.695-.892-1.206-1.918-1.533-3.076-.327-1.169-.455-2.425-.385-3.768Zm4.934-.1-.019.358c-.04.774-.01 1.5.092 2.178.101.679.286 1.282.555 1.809.281.517.659.93 1.133 1.24.474.31 1.064.484 1.77.521.682.036 1.276-.076 1.78-.335.504-.258.917-.63 1.239-1.115.323-.497.57-1.077.742-1.741.183-.664.295-1.383.335-2.156l.019-.359c.039-.751.002-1.46-.111-2.127-.101-.679-.292-1.282-.573-1.81-.268-.539-.639-.969-1.113-1.291-.474-.321-1.057-.5-1.751-.537-.695-.036-1.294.081-1.799.351-.493.271-.907.66-1.242 1.167-.322.497-.569 1.077-.741 1.741-.171.653-.277 1.355-.316 2.106Zm23.416-15.63-1.374 26.214-4.932-.258 1.373-26.214 4.933.258Zm10.903 23.264 4.876-14.667 5.137.269-7.197 18.139-3.106-.162.29-3.579Zm-2.036-15.03 3.333 15.115-.102 3.571-3.123-.163-5.262-18.793 5.154.27Zm23.273 20.079c-1.434-.075-2.707-.37-3.821-.885-1.114-.526-2.047-1.219-2.8-2.08-.743-.861-1.296-1.842-1.66-2.945-.364-1.115-.513-2.286-.449-3.515l.036-.683c.073-1.399.339-2.669.798-3.808.459-1.14 1.081-2.117 1.865-2.932.796-.814 1.74-1.42 2.834-1.819 1.094-.411 2.312-.581 3.654-.51 1.309.068 2.458.345 3.448.831.99.485 1.805 1.144 2.446 1.976.653.833 1.126 1.816 1.42 2.95.295 1.122.407 2.36.336 3.714l-.107 2.048-14.677-.769.171-3.277 9.848.516.019-.376c.036-.682-.057-1.298-.279-1.845-.211-.559-.552-1.01-1.025-1.355-.472-.344-1.09-.536-1.852-.576-.649-.034-1.213.079-1.695.339-.481.26-.883.638-1.206 1.135-.311.497-.559 1.089-.743 1.775-.173.675-.28 1.423-.323 2.242l-.036.683c-.039.739.028 1.427.2 2.064.183.637.457 1.193.82 1.668.374.476.833.854 1.378 1.134.555.28 1.192.439 1.908.476.888.047 1.722-.081 2.502-.382.792-.312 1.485-.807 2.08-1.483l2.253 2.72c-.416.537-.973 1.044-1.671 1.521-.687.477-1.505.856-2.456 1.137-.949.27-2.021.374-3.216.311Zm-75.938 31.404 2.804.147c-.146 1.064-.495 2.002-1.047 2.812-.552.804-1.301 1.421-2.245 1.851-.945.43-2.075.611-3.391.542-1.01-.053-1.909-.28-2.697-.681-.788-.409-1.453-.961-1.994-1.657-.54-.703-.939-1.537-1.196-2.503-.249-.965-.343-2.034-.281-3.208l.071-1.357c.061-1.174.27-2.228.626-3.161.364-.933.855-1.721 1.474-2.363.619-.65 1.345-1.133 2.179-1.449.841-.316 1.774-.447 2.798-.394 1.301.069 2.388.365 3.26.891.872.525 1.532 1.224 1.982 2.094.456.871.712 1.856.767 2.953l-2.804-.147c-.038-.706-.171-1.317-.399-1.831-.22-.514-.566-.915-1.039-1.202-.472-.294-1.1-.462-1.885-.503-.643-.034-1.21.056-1.701.27-.491.215-.91.545-1.255.992-.346.446-.619 1.006-.819 1.677-.193.665-.311 1.431-.357 2.298l-.072 1.38c-.043.822-.011 1.574.095 2.254.115.673.309 1.257.583 1.751.282.495.651.885 1.108 1.172.458.286 1.015.446 1.673.481.8.042 1.454-.051 1.961-.28.514-.228.911-.578 1.191-1.051.287-.479.491-1.072.61-1.778Zm5.757-.317.013-.258c.046-.875.216-1.679.509-2.414.294-.742.694-1.377 1.202-1.905.515-.535 1.129-.938 1.84-1.208.719-.277 1.52-.393 2.402-.346.89.046 1.674.245 2.353.595.687.344 1.259.808 1.715 1.395.457.578.793 1.252 1.008 2.021.215.761.299 1.578.253 2.453l-.013.258c-.046.875-.216 1.679-.509 2.414-.293.734-.698 1.369-1.213 1.904-.515.528-1.128.931-1.84 1.208-.712.27-1.509.382-2.391.336-.889-.046-1.678-.241-2.364-.585-.679-.35-1.247-.815-1.704-1.394-.457-.586-.793-1.26-1.008-2.021-.215-.761-.299-1.578-.253-2.453Zm2.716-.116-.014.257c-.028.546.001 1.065.087 1.557.087.492.241.927.461 1.306.22.379.511.683.874.912.363.229.802.357 1.318.384.501.026.94-.056 1.318-.246.384-.19.706-.462.964-.815.259-.354.457-.771.595-1.251.145-.48.231-.993.26-1.539l.014-.257c.028-.539-.005-1.05-.1-1.535-.087-.492-.244-.931-.471-1.318-.219-.386-.511-.694-.873-.923-.355-.236-.787-.367-1.296-.394-.508-.027-.955.059-1.34.256-.377.19-.695.466-.954.827-.259.361-.457.782-.595 1.262-.137.473-.22.978-.248 1.517Zm15.198-5.164-.636 12.134-2.714-.142.636-12.134 2.714.142Zm-2.726-3.337c.021-.411.174-.744.457-.999.291-.263.68-.381 1.166-.355.478.025.848.183 1.111.474.262.284.383.631.361 1.042-.021.404-.177.733-.468.988-.291.255-.675.369-1.154.344-.486-.025-.86-.18-1.122-.463-.255-.284-.372-.627-.351-1.031Zm9.38 6.283-.5 9.544-2.703-.141.636-12.135 2.546.133.021 2.599Zm-.641 3.003-.874-.057c.053-.859.213-1.642.483-2.348.277-.705.638-1.304 1.084-1.798.453-.494.982-.864 1.587-1.109.606-.253 1.271-.361 1.997-.323.583.031 1.106.14 1.568.33.47.182.864.461 1.182.838.325.376.562.857.711 1.443.15.577.203 1.281.16 2.111l-.411 7.839-2.714-.142.411-7.851c.031-.583-.031-1.047-.185-1.393-.147-.352-.377-.608-.691-.767-.306-.166-.695-.261-1.166-.286-.464-.024-.884.051-1.26.227-.377.175-.701.424-.973.747-.264.324-.475.702-.633 1.137-.158.434-.25.901-.276 1.402Zm23.275 7.754-3.487-.183.139-2.231 3.465.182c1.01.053 1.866-.123 2.57-.529.711-.405 1.26-1.01 1.647-1.814.394-.804.621-1.774.681-2.911l.046-.886c.046-.882-.014-1.669-.18-2.36-.159-.69-.42-1.278-.785-1.761-.357-.484-.809-.86-1.357-1.129-.541-.268-1.174-.421-1.899-.459l-3.657-.192.118-2.243 3.656.192c1.084.057 2.065.292 2.943.705.879.406 1.625.966 2.24 1.681.622.715 1.084 1.556 1.386 2.524.302.968.422 2.039.36 3.213l-.045.863c-.062 1.174-.293 2.227-.694 3.158-.401.931-.949 1.719-1.642 2.365-.693.639-1.506 1.117-2.437 1.436-.924.311-1.946.437-3.068.379Zm-1.073-16.431-.856 16.33-2.815-.148.856-16.329 2.815.147Zm17.552 17.519c-.897-.047-1.7-.235-2.41-.565-.703-.336-1.294-.783-1.774-1.341-.474-.557-.826-1.202-1.057-1.933-.232-.732-.326-1.513-.283-2.343l.024-.449c.05-.949.233-1.802.55-2.557.317-.756.737-1.394 1.259-1.914.522-.527 1.127-.919 1.816-1.175.688-.257 1.424-.364 2.21-.323.867.045 1.618.231 2.253.557.635.325 1.152.765 1.551 1.318.406.546.695 1.187.866 1.923.179.737.245 1.543.199 2.417l-.06 1.156-9.376-.492.102-1.94 6.706.351.012-.213c.01-.486-.063-.948-.22-1.383-.15-.435-.397-.793-.743-1.074-.345-.28-.816-.436-1.415-.467-.448-.024-.853.052-1.215.228-.354.169-.656.427-.906.774-.251.346-.453.771-.607 1.272-.146.495-.235 1.057-.268 1.685l-.024.448c-.027.531.018 1.028.136 1.492.126.456.322.86.589 1.211.266.352.596.635.99.851.394.208.849.325 1.365.352.65.034 1.236-.066 1.758-.301.523-.235.983-.582 1.382-1.041l1.352 1.454c-.282.368-.641.716-1.078 1.046-.437.322-.964.575-1.581.76-.609.186-1.31.258-2.103.216Zm11.306-9.202-.501 9.544-2.702-.142.636-12.135 2.545.134.022 2.599Zm-.641 3.002-.875-.057c.053-.859.214-1.642.483-2.347.277-.705.639-1.305 1.084-1.799.454-.494.983-.863 1.588-1.109.605-.253 1.271-.361 1.996-.323.583.031 1.106.141 1.568.33.471.182.865.461 1.182.838.325.377.563.858.712 1.443.15.578.203 1.281.159 2.111l-.411 7.84-2.714-.143.412-7.85c.03-.583-.032-1.048-.186-1.393-.146-.353-.377-.608-.691-.767-.306-.166-.695-.262-1.166-.286-.463-.025-.883.051-1.26.226-.376.175-.7.424-.972.747-.265.324-.476.703-.633 1.137-.158.434-.25.901-.276 1.402Zm11.189 1.172.013-.258c.046-.875.216-1.68.509-2.414.294-.742.694-1.377 1.202-1.905.515-.535 1.129-.938 1.84-1.208.719-.277 1.52-.393 2.402-.347.89.047 1.674.246 2.353.596.687.343 1.259.808 1.715 1.394.457.579.793 1.253 1.008 2.021.215.761.299 1.579.254 2.454l-.014.258c-.046.875-.215 1.679-.509 2.414-.293.734-.698 1.369-1.213 1.904-.515.528-1.128.93-1.84 1.208-.712.27-1.508.382-2.391.336-.889-.047-1.678-.242-2.364-.585-.679-.351-1.247-.815-1.704-1.394-.457-.586-.793-1.26-1.008-2.021-.215-.761-.299-1.579-.253-2.453Zm2.716-.117-.013.258c-.029.546 0 1.065.087 1.557.086.492.24.927.46 1.306.22.379.511.683.874.912.363.229.802.357 1.318.384.501.026.94-.056 1.318-.246.384-.19.706-.462.965-.816.258-.353.456-.77.594-1.251.145-.479.232-.992.26-1.538l.014-.258c.028-.538-.005-1.05-.1-1.534-.086-.492-.243-.932-.471-1.318-.219-.387-.51-.694-.873-.923-.355-.236-.787-.368-1.295-.394-.509-.027-.956.058-1.341.256-.377.19-.695.465-.954.827-.259.361-.457.782-.595 1.262-.137.472-.22.978-.248 1.516Zm14.878-2.706-.506 9.667-2.703-.141.636-12.135 2.546.133.027 2.476Zm-.602 3.128-.919-.059c.044-.838.189-1.606.436-2.305.246-.7.585-1.3 1.016-1.802.432-.51.954-.891 1.567-1.144.621-.26 1.328-.369 2.12-.328.553.029 1.054.138 1.501.326.456.182.843.453 1.161.815.326.362.564.817.716 1.364.158.548.218 1.204.178 1.966l-.424 8.086-2.703-.141.412-7.851c.031-.59-.035-1.059-.197-1.405-.154-.345-.392-.597-.713-.757-.314-.166-.695-.261-1.144-.285-.508-.026-.947.048-1.316.224-.362.176-.663.426-.905.751s-.427.705-.555 1.14c-.128.436-.205.904-.231 1.405Zm7.563-.323-1.282.214c.038-.733.175-1.419.411-2.059.244-.647.577-1.207.999-1.68.43-.479.948-.846 1.554-1.099.605-.253 1.289-.36 2.052-.32.62.033 1.169.148 1.646.345.485.191.89.474 1.215.851.325.377.563.858.712 1.443.149.578.203 1.27.161 2.078l-.412 7.85-2.714-.142.412-7.862c.032-.613-.032-1.092-.194-1.438-.155-.346-.389-.59-.704-.734-.314-.152-.696-.239-1.145-.263-.418-.022-.792.038-1.122.178-.322.133-.599.332-.83.597-.231.258-.412.56-.542.905-.123.346-.196.725-.217 1.136Zm14.516-4.154-.636 12.135-2.714-.142.636-12.135 2.714.142Zm-2.727-3.337c.022-.411.174-.744.457-.999.291-.262.68-.38 1.166-.355.478.025.849.183 1.111.474.262.284.383.631.361 1.043-.021.403-.177.733-.467.987-.291.255-.676.37-1.154.345-.486-.026-.861-.18-1.123-.464-.255-.283-.372-.627-.351-1.031Zm9.38 6.284-.5 9.544-2.703-.142.636-12.135 2.546.134.021 2.599Zm-.641 3.003-.874-.058c.053-.859.214-1.642.483-2.347.277-.705.638-1.305 1.084-1.799.453-.493.982-.863 1.588-1.109.605-.253 1.271-.36 1.996-.322.583.03 1.106.14 1.568.329.47.182.864.462 1.182.838.325.377.562.858.712 1.443.149.578.202 1.281.159 2.111l-.411 7.84-2.714-.143.411-7.85c.031-.583-.031-1.048-.185-1.393-.147-.353-.377-.608-.691-.767-.306-.166-.695-.262-1.166-.286-.463-.025-.883.051-1.26.226-.376.175-.701.424-.973.748-.264.323-.475.702-.633 1.136-.157.434-.249.902-.276 1.403Zm18.363 5.055.303-5.787c.023-.433-.036-.811-.176-1.133-.141-.323-.367-.578-.68-.767-.305-.188-.697-.295-1.175-.32-.441-.023-.826.031-1.156.164-.329.133-.59.321-.783.566-.193.245-.298.528-.314.85l-2.692-.141c.025-.479.165-.936.421-1.373.255-.436.612-.819 1.072-1.147.459-.328 1.001-.577 1.625-.747.623-.17 1.316-.235 2.079-.195.912.048 1.712.244 2.398.587.695.344 1.227.837 1.599 1.478.379.635.544 1.416.495 2.343l-.282 5.394c-.029.554-.018 1.053.034 1.498.059.438.155.822.288 1.151l-.01.179-2.77-.145c-.111-.298-.193-.673-.244-1.126-.044-.46-.054-.902-.032-1.329Zm.652-4.925-.065 1.672-1.94-.101c-.501-.027-.945-.001-1.331.076-.386.07-.711.188-.975.354-.264.166-.466.373-.606.62-.14.248-.219.532-.236.854-.017.321.043.621.178.898.136.269.342.486.618.651.284.164.631.258 1.043.279.553.029 1.041-.058 1.464-.261.431-.209.774-.472 1.031-.789.257-.324.401-.635.431-.933l.812 1.246c-.106.301-.276.622-.511.962-.236.341-.537.662-.906.965-.36.296-.789.532-1.285.709-.489.177-1.048.248-1.676.216-.792-.042-1.491-.236-2.095-.583-.603-.354-1.067-.809-1.39-1.366-.323-.564-.466-1.19-.43-1.878.033-.643.183-1.205.448-1.686.273-.488.654-.888 1.142-1.199.497-.312 1.093-.535 1.79-.671.697-.144 1.487-.192 2.369-.146l2.12.111Zm12.108-4.134-.104 1.974-6.841-.358.104-1.974 6.841.358Zm-4.712-3.227 2.703.142-.616 11.753c-.019.374.018.665.112.872.102.2.248.339.439.417.191.077.417.123.679.137.187.009.367.008.54-.006.173-.013.313-.028.419-.045l-.097 2.064c-.228.055-.493.101-.795.138-.294.037-.632.046-1.013.026-.621-.033-1.165-.17-1.632-.412-.467-.249-.821-.632-1.064-1.147-.243-.515-.343-1.187-.299-2.017l.624-11.922Zm10.84 3.549-.636 12.134-2.714-.142.636-12.135 2.714.143Zm-2.727-3.337c.022-.411.174-.745.458-1 .291-.262.679-.38 1.165-.355.479.025.849.184 1.111.475.263.283.383.631.362 1.042-.022.404-.177.733-.468.988-.291.254-.676.369-1.154.344-.486-.025-.86-.18-1.123-.464-.255-.283-.372-.626-.351-1.03Zm5.995 9.727.014-.258c.045-.875.215-1.68.509-2.414.293-.742.694-1.377 1.201-1.905.516-.535 1.129-.938 1.841-1.208.719-.277 1.52-.393 2.402-.346.89.046 1.674.245 2.353.595.687.344 1.258.808 1.715 1.395.457.578.793 1.252 1.007 2.02.215.761.3 1.579.254 2.454l-.013.258c-.046.875-.216 1.679-.509 2.414-.294.734-.698 1.369-1.214 1.904-.514.528-1.128.931-1.84 1.208-.711.27-1.508.382-2.39.336-.89-.047-1.678-.242-2.365-.585-.679-.35-1.247-.815-1.704-1.394-.456-.586-.792-1.26-1.007-2.021-.215-.761-.3-1.578-.254-2.453Zm2.716-.117-.013.258c-.029.546 0 1.065.087 1.557s.24.927.46 1.306c.22.379.512.683.875.912.362.229.802.357 1.318.384.501.026.94-.056 1.317-.246.385-.19.707-.462.965-.815.258-.354.456-.771.594-1.251.145-.48.232-.993.261-1.539l.013-.258c.028-.538-.005-1.049-.099-1.534-.087-.492-.244-.931-.471-1.318-.22-.387-.511-.694-.874-.923-.355-.236-.787-.367-1.295-.394-.508-.027-.955.059-1.34.256-.378.19-.696.466-.955.827-.258.361-.457.782-.594 1.262-.138.473-.22.978-.249 1.516Zm14.85-2.584-.5 9.544-2.703-.141.636-12.135 2.546.133.021 2.599Zm-.641 3.003-.874-.057c.052-.859.213-1.642.483-2.348.277-.705.638-1.304 1.084-1.798.453-.494.982-.864 1.587-1.109.606-.253 1.271-.361 1.996-.323.584.031 1.106.14 1.569.33.47.182.864.461 1.182.838.325.376.562.857.711 1.443.15.577.203 1.281.16 2.111l-.411 7.839-2.714-.142.411-7.851c.031-.583-.031-1.047-.186-1.393-.146-.352-.376-.608-.69-.767-.307-.166-.695-.261-1.166-.286-.464-.024-.884.051-1.26.227-.377.175-.701.424-.973.747-.264.324-.475.702-.633 1.136-.158.435-.25.902-.276 1.403Z"
-      />
-      <rect
-        width="200"
-        height="13"
-        x="493.872"
-        y="366.058"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 493.872 366.058)"
-      />
-      <rect
-        width="300"
-        height="13"
-        x="492.302"
-        y="396.017"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 492.302 396.017)"
-      />
-      <rect
-        width="230"
-        height="13"
-        x="490.732"
-        y="425.975"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 490.732 425.975)"
-      />
-      <rect
-        width="100"
-        height="13"
-        x="711.573"
-        y="377.467"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 711.573 377.467)"
-      />
-      <rect
-        width="100"
-        height="13"
-        x="874.205"
-        y="446.072"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 874.205 446.072)"
-      />
-      <rect
-        width="150"
-        height="13"
-        x="825.417"
-        y="383.433"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 825.417 383.433)"
-      />
-      <rect
-        width="176"
-        height="13"
-        x="808.867"
-        y="412.607"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 808.867 412.607)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="732.4"
-        y="438.641"
-        fill="#C1D3D8"
-        rx="5"
-        transform="rotate(3 732.4 438.641)"
-      />
-      <path
-        fill="#D1DBDD"
-        d="m489.357 456.43 283.61079 14.86341-10.88588 207.71494-283.61079-14.8634zm291.6 15.282 200.72454 10.51953-10.88588 207.71494-200.72454-10.51953z"
-      />
-      <path
-        fill="#434040"
-        d="m505.953 480.703 2.438.127c-.127.926-.43 1.741-.91 2.446-.48.699-1.131 1.235-1.952 1.609-.822.375-1.805.532-2.949.472-.878-.046-1.66-.244-2.345-.593-.686-.355-1.264-.835-1.734-1.44-.47-.611-.817-1.337-1.04-2.177-.217-.839-.298-1.769-.245-2.79l.062-1.18c.054-1.02.235-1.937.545-2.748.316-.812.743-1.497 1.281-2.055.538-.565 1.17-.985 1.895-1.26.731-.275 1.543-.389 2.433-.342 1.132.059 2.076.317 2.835.774.758.457 1.333 1.064 1.723 1.822.397.757.62 1.613.667 2.567l-2.438-.127c-.033-.615-.148-1.146-.346-1.593-.192-.447-.493-.795-.904-1.045-.41-.256-.957-.402-1.639-.438-.56-.029-1.053.049-1.48.236-.427.186-.79.473-1.091.862-.301.388-.538.874-.712 1.458-.167.578-.271 1.245-.31 1.999l-.063 1.199c-.038.715-.01 1.369.083 1.96.1.586.269 1.093.507 1.523.245.43.566.77.963 1.019.398.249.883.388 1.455.418.696.037 1.264-.044 1.705-.243.447-.198.792-.503 1.035-.914.25-.417.427-.932.531-1.546Zm5.006-.276.012-.224c.039-.761.187-1.46.442-2.099.255-.645.604-1.197 1.045-1.657.448-.465.982-.815 1.6-1.05.626-.241 1.322-.342 2.089-.301.774.04 1.456.213 2.046.518.597.298 1.095.702 1.492 1.212.397.503.689 1.089.876 1.757.187.662.26 1.373.22 2.134l-.011.224c-.04.761-.188 1.461-.443 2.099-.255.639-.607 1.191-1.055 1.656-.448.459-.981.809-1.6 1.051-.619.235-1.312.332-2.079.292-.773-.041-1.459-.21-2.056-.509-.59-.305-1.084-.709-1.482-1.212-.397-.51-.689-1.096-.876-1.757-.187-.662-.26-1.373-.22-2.134Zm2.362-.101-.012.224c-.025.475 0 .926.076 1.354.075.427.209.806.4 1.136.191.329.445.593.76.792.316.199.698.311 1.146.334.436.023.818-.048 1.146-.214.335-.165.614-.401.839-.709.225-.307.397-.67.517-1.087.126-.417.201-.863.226-1.338l.012-.224c.025-.468-.004-.913-.087-1.335-.075-.427-.211-.809-.409-1.146-.191-.336-.444-.603-.76-.802-.308-.206-.684-.32-1.126-.343-.442-.023-.831.051-1.165.223-.329.165-.605.405-.83.719-.225.314-.398.68-.517 1.097-.12.411-.192.851-.216 1.319Zm16.698 4.083.67-12.795 2.36.123-.785 14.98-2.135-.112-.11-2.196Zm-6.71-3.335.011-.204c.042-.8.174-1.523.397-2.17.223-.653.526-1.208.909-1.664.382-.462.837-.807 1.364-1.033.527-.233 1.113-.333 1.757-.299.637.033 1.189.186 1.658.458.468.272.859.648 1.173 1.127.314.473.555 1.034.723 1.681.169.641.275 1.351.319 2.129l-.034.654c-.124.749-.304 1.431-.538 2.045-.234.613-.532 1.139-.894 1.576-.362.437-.793.767-1.294.988-.494.222-1.063.316-1.706.283-.638-.034-1.206-.197-1.706-.491-.493-.293-.905-.689-1.238-1.189-.332-.5-.576-1.08-.73-1.74-.154-.667-.211-1.383-.171-2.151Zm2.361-.081-.011.205c-.025.481-.006.931.057 1.352.069.421.189.796.361 1.124.172.322.403.582.693.779.296.192.656.299 1.078.321.533.028.978-.066 1.335-.282.357-.216.643-.517.859-.904.223-.392.383-.834.48-1.324l.093-1.765c-.019-.386-.082-.748-.188-1.086-.099-.337-.244-.635-.432-.893-.188-.264-.429-.472-.72-.624-.286-.158-.63-.248-1.033-.269-.429-.023-.798.049-1.106.215-.309.167-.569.404-.781.713-.205.308-.364.672-.477 1.089-.114.418-.183.868-.208 1.349Zm14.692 6.344c-.78-.041-1.479-.205-2.096-.491-.611-.293-1.125-.682-1.543-1.167-.411-.484-.718-1.045-.919-1.681-.201-.636-.283-1.316-.245-2.037l.02-.39c.043-.826.203-1.567.479-2.224.275-.657.64-1.212 1.094-1.664.454-.459.981-.799 1.579-1.022.598-.223 1.239-.317 1.922-.281.754.039 1.407.201 1.959.484.553.283 1.002.665 1.349 1.146.353.475.604 1.033.753 1.673.155.64.213 1.341.173 2.102l-.052 1.004-8.153-.427.088-1.687 5.832.305.01-.185c.009-.423-.055-.824-.191-1.203-.131-.378-.346-.689-.646-.933-.3-.244-.71-.38-1.23-.407-.39-.021-.743.046-1.057.199-.308.147-.57.371-.788.672-.218.302-.394.671-.528 1.107-.126.43-.204.918-.233 1.464l-.02.391c-.024.461.015.893.118 1.297.109.396.28.748.512 1.053.231.306.518.552.86.74.343.181.739.283 1.187.306.566.03 1.076-.057 1.53-.262.454-.204.854-.506 1.201-.905l1.176 1.264c-.245.32-.558.623-.938.91-.38.28-.838.5-1.374.661-.53.161-1.14.224-1.829.188Zm260.277 5.934-.041.781c-.057 1.072-.247 2.027-.571 2.864-.324.837-.762 1.541-1.313 2.112-.545.571-1.187.997-1.925 1.278-.739.274-1.546.389-2.424.343-.871-.046-1.66-.244-2.365-.594-.698-.356-1.295-.847-1.79-1.472-.496-.626-.867-1.373-1.115-2.24-.242-.866-.334-1.836-.278-2.909l.041-.78c.056-1.073.249-2.024.58-2.854.33-.837.771-1.541 1.323-2.112.558-.577 1.203-1.003 1.935-1.277.738-.281 1.543-.399 2.414-.353.878.046 1.669.247 2.374.604.705.35 1.302.841 1.791 1.472.488.626.85 1.372 1.085 2.238.242.86.335 1.827.279 2.899Zm-2.489.652.042-.799c.041-.794 0-1.497-.124-2.109-.118-.619-.315-1.141-.593-1.567-.271-.431-.615-.762-1.034-.993-.417-.237-.896-.369-1.436-.398-.539-.028-1.026.054-1.46.247-.433.185-.811.479-1.132.879-.314.394-.565.893-.753 1.496-.187.596-.302 1.291-.344 2.084l-.042.8c-.041.793 0 1.499.124 2.119.124.619.325 1.148.602 1.586.283.432.631.767 1.042 1.003.412.231.888.36 1.427.388.546.029 1.036-.05 1.47-.236.433-.192.808-.489 1.122-.89.315-.407.563-.912.744-1.515.188-.603.303-1.302.345-2.095Zm11.787 4.852.423-8.065 2.36.124-.553 10.552-2.223-.117-.007-2.494Zm.447-2.176.791.021c-.037.709-.149 1.358-.337 1.948-.187.584-.455 1.088-.803 1.513-.348.419-.782.738-1.302.959-.519.214-1.134.302-1.842.265-.514-.027-.981-.126-1.403-.298-.421-.172-.779-.423-1.075-.751-.289-.328-.505-.747-.648-1.256-.143-.51-.196-1.113-.16-1.808l.358-6.817 2.35.123-.358 6.837c-.02.383.008.707.086.972.077.258.19.47.338.634.148.164.324.284.529.36.204.076.424.12.658.132.669.036 1.203-.067 1.6-.307.404-.246.699-.589.885-1.029.193-.44.304-.939.333-1.498Zm10.581-5.333-.09 1.716-5.949-.311.09-1.717 5.949.312Zm-4.097-2.806 2.35.123-.536 10.22c-.017.325.016.578.097.758.089.175.216.295.382.363.166.067.363.107.591.119.162.008.319.007.469-.005.151-.012.272-.025.364-.04l-.084 1.795c-.198.049-.428.089-.691.121-.256.032-.55.039-.881.022-.54-.028-1.013-.148-1.419-.358-.406-.217-.714-.549-.926-.997-.211-.448-.297-1.033-.26-1.754l.544-10.367Zm9.163 5.106-.659 12.58-2.351-.123.766-14.609 2.165.113.079 2.039Zm6.71 3.51-.011.205c-.04.767-.168 1.474-.385 2.121-.209.641-.502 1.196-.879 1.665-.37.463-.815.815-1.336 1.055-.521.24-1.113.343-1.777.308-.656-.035-1.225-.185-1.707-.452-.474-.272-.869-.642-1.184-1.108-.314-.466-.56-1.007-.736-1.623-.17-.621-.277-1.299-.324-2.031l.042-.79c.125-.769.304-1.464.539-2.084.241-.62.543-1.148.905-1.586.362-.437.793-.767 1.293-.988.501-.222 1.076-.315 1.726-.281.664.034 1.245.195 1.745.482.5.281.914.668 1.24 1.161.326.486.56 1.065.701 1.738.141.666.19 1.402.148 2.208Zm-2.361.082.011-.205c.025-.488.004-.942-.066-1.363-.068-.427-.192-.805-.37-1.134-.179-.329-.416-.589-.712-.781-.29-.197-.646-.307-1.069-.33-.416-.021-.777.031-1.084.159-.306.121-.566.299-.781.536-.214.236-.385.517-.513.843-.128.32-.224.67-.29 1.051l-.099 1.892c.053.472.164.908.332 1.308s.418.726.751.978c.339.246.782.384 1.328.412.422.022.788-.05 1.096-.216.309-.167.566-.404.771-.713.212-.315.375-.681.488-1.099.113-.418.182-.864.207-1.338Zm11.556 3.294.423-8.065 2.36.124-.553 10.552-2.223-.117-.007-2.494Zm.447-2.176.791.021c-.037.709-.149 1.358-.337 1.948-.187.584-.454 1.088-.803 1.513-.348.419-.782.738-1.302.959-.519.214-1.134.302-1.842.265-.514-.027-.981-.126-1.403-.298-.421-.172-.779-.423-1.075-.751-.289-.328-.505-.747-.648-1.256-.143-.51-.196-1.113-.16-1.808l.358-6.817 2.35.123-.358 6.837c-.02.383.008.707.086.972.077.258.19.47.338.634.148.164.324.284.529.36.204.076.424.12.658.132.669.036 1.203-.067 1.6-.307.404-.246.699-.589.885-1.029.193-.44.304-.939.333-1.498Zm10.581-5.333-.09 1.716-5.949-.311.09-1.717 5.949.312Zm-4.097-2.806 2.35.123-.536 10.22c-.017.325.016.578.098.758.088.175.216.295.382.363.165.067.362.107.59.119.162.008.319.007.469-.005.151-.012.272-.025.364-.04l-.084 1.795c-.198.049-.428.089-.691.121-.256.032-.549.039-.881.022-.54-.028-1.013-.148-1.419-.358-.406-.217-.714-.549-.925-.997-.212-.448-.298-1.033-.261-1.754l.544-10.367Z"
-      />
-      <rect
-        width="70"
-        height="10"
-        x="495.253"
-        y="496.794"
-        fill="#98C7B9"
-        rx="5"
-        transform="rotate(3 495.253 496.794)"
-      />
-      <rect
-        width="90"
-        height="10"
-        x="553.282"
-        y="516.858"
-        fill="#F8B1B1"
-        rx="5"
-        transform="rotate(3 553.282 516.858)"
-      />
-      <rect
-        width="50"
-        height="10"
-        x="523.432"
-        y="532.317"
-        fill="#84A78C"
-        rx="5"
-        transform="rotate(3 523.432 532.317)"
-      />
-      <rect
-        width="50"
-        height="10"
-        x="559.492"
-        y="551.23"
-        fill="#98C7B9"
-        rx="5"
-        transform="rotate(3 559.492 551.23)"
-      />
-      <rect
-        width="70"
-        height="10"
-        x="618.411"
-        y="554.318"
-        fill="#5982D2"
-        rx="5"
-        transform="rotate(3 618.411 554.318)"
-      />
-      <rect
-        width="30"
-        height="10"
-        x="522.542"
-        y="549.294"
-        fill="#AEADF2"
-        rx="5"
-        transform="rotate(3 522.542 549.294)"
-      />
-      <rect
-        width="45"
-        height="10"
-        x="521.653"
-        y="566.27"
-        fill="#DA9FE9"
-        rx="5"
-        transform="rotate(3 521.653 566.27)"
-      />
-      <rect
-        width="68"
-        height="10"
-        x="500.79"
-        y="582.2"
-        fill="#F8B1B1"
-        rx="5"
-        transform="rotate(3 500.79 582.2)"
-      />
-      <rect
-        width="35"
-        height="10"
-        x="519.873"
-        y="600.224"
-        fill="#DA9FE9"
-        rx="5"
-        transform="rotate(3 519.873 600.224)"
-      />
-      <rect
-        width="78"
-        height="10"
-        x="562.814"
-        y="602.474"
-        fill="#5982D2"
-        rx="5"
-        transform="rotate(3 562.814 602.474)"
-      />
-      <rect
-        width="45"
-        height="10"
-        x="648.696"
-        y="606.975"
-        fill="#3EABE8"
-        rx="5"
-        transform="rotate(3 648.696 606.975)"
-      />
-      <rect
-        width="65"
-        height="10"
-        x="518.983"
-        y="617.201"
-        fill="#AEADF2"
-        rx="5"
-        transform="rotate(3 518.983 617.201)"
-      />
-      <rect
-        width="54"
-        height="10"
-        x="500.118"
-        y="633.235"
-        fill="#8ECCD5"
-        rx="5"
-        transform="rotate(3 500.118 633.235)"
-      />
-      <rect
-        width="34"
-        height="10"
-        x="560.036"
-        y="636.375"
-        fill="#84A78C"
-        rx="5"
-        transform="rotate(3 560.036 636.375)"
-      />
-      <rect
-        width="46"
-        height="10"
-        x="588.887"
-        y="620.864"
-        fill="#3EABE8"
-        rx="5"
-        transform="rotate(3 588.887 620.864)"
-      />
-      <rect
-        width="25"
-        height="10"
-        x="639.818"
-        y="623.533"
-        fill="#98C7B9"
-        rx="5"
-        transform="rotate(3 639.818 623.533)"
-      />
-      <rect
-        width="90"
-        height="10"
-        x="576.686"
-        y="586.178"
-        fill="#98C7B9"
-        rx="5"
-        transform="rotate(3 576.686 586.178)"
-      />
-      <rect
-        width="45"
-        height="10"
-        x="572.583"
-        y="568.94"
-        fill="#8ECCD5"
-        rx="5"
-        transform="rotate(3 572.583 568.94)"
-      />
-      <rect
-        width="77"
-        height="10"
-        x="625.51"
-        y="571.713"
-        fill="#84A78C"
-        rx="5"
-        transform="rotate(3 625.51 571.713)"
-      />
-      <rect
-        width="70"
-        height="10"
-        x="579.355"
-        y="535.248"
-        fill="#DA9FE9"
-        rx="5"
-        transform="rotate(3 579.355 535.248)"
-      />
-      <rect
-        width="66"
-        height="10"
-        x="655.251"
-        y="539.225"
-        fill="#8ECCD5"
-        rx="5"
-        transform="rotate(3 655.251 539.225)"
-      />
-      <rect
-        width="54"
-        height="10"
-        x="650.149"
-        y="521.935"
-        fill="#AEADF2"
-        rx="5"
-        transform="rotate(3 650.149 521.935)"
-      />
-      <rect
-        width="40"
-        height="10"
-        x="506.346"
-        y="514.398"
-        fill="#DA9FE9"
-        rx="5"
-        transform="rotate(3 506.346 514.398)"
-      />
-      <rect
-        width="184"
-        height="30"
-        x="780.415"
-        y="634.907"
-        fill="#929899"
-        rx="5"
-        transform="rotate(3 780.415 634.907)"
-      />
-      <rect
-        width="183"
-        height="29"
-        x="780.888"
-        y="635.433"
-        stroke="#68888F"
-        stroke-opacity=".8"
-        rx="4.5"
-        transform="rotate(3 780.888 635.433)"
-      />
-      <circle
-        cx="796.607"
-        cy="650.776"
-        r="10"
-        fill="#F24E1E"
-        transform="rotate(3 796.607 650.776)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="818.943"
-        y="644.937"
-        fill="#F24E1E"
-        rx="5"
-        transform="rotate(3 818.943 644.937)"
-      />
-      <path
-        stroke="#fff"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        stroke-width="1.25"
-        d="M799.882 647.416 792 654.931m.393-7.931 7.097 8.346"
-      />
-      <rect
-        width="184"
-        height="30"
-        x="782.508"
-        y="594.962"
-        fill="#929899"
-        rx="5"
-        transform="rotate(3 782.508 594.962)"
-      />
-      <rect
-        width="183"
-        height="29"
-        x="782.982"
-        y="595.487"
-        stroke="#68888F"
-        stroke-opacity=".8"
-        rx="4.5"
-        transform="rotate(3 782.982 595.487)"
-      />
-      <circle
-        cx="798.7"
-        cy="610.831"
-        r="10"
-        fill="#F24E1E"
-        transform="rotate(3 798.7 610.831)"
-      />
-      <path
-        stroke="#fff"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        stroke-width="1.25"
-        d="M802.882 607.416 795 614.931m.393-7.931 7.097 8.346"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="821.036"
-        y="604.992"
-        fill="#F24E1E"
-        rx="5"
-        transform="rotate(3 821.036 604.992)"
-      />
-      <rect
-        width="184"
-        height="30"
-        x="784.602"
-        y="555.017"
-        fill="#929899"
-        rx="5"
-        transform="rotate(3 784.602 555.017)"
-      />
-      <rect
-        width="183"
-        height="29"
-        x="785.075"
-        y="555.542"
-        stroke="#68888F"
-        stroke-opacity=".8"
-        rx="4.5"
-        transform="rotate(3 785.075 555.542)"
-      />
-      <circle
-        cx="800.794"
-        cy="570.886"
-        r="10"
-        fill="#4ECB71"
-        transform="rotate(3 800.794 570.886)"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="823.13"
-        y="565.047"
-        fill="#4ECB71"
-        rx="5"
-        transform="rotate(3 823.13 565.047)"
-      />
-      <path
-        fill="#fff"
-        fill-rule="evenodd"
-        d="M807.134 566.904c.085.077.153.169.201.27.048.101.074.21.077.321.004.11-.016.22-.057.323-.042.103-.104.197-.184.276l-7.475 7.445c-.084.083-.185.149-.297.194-.113.044-.234.065-.356.063-.122-.002-.243-.029-.355-.077-.111-.049-.211-.118-.293-.205l-3.94-4.149c-.154-.167-.233-.384-.22-.604.014-.22.117-.426.289-.572.172-.147.399-.223.631-.211.233.011.452.108.612.27l3.296 3.469 6.818-6.79c.161-.161.384-.253.619-.257.235-.004.463.079.634.234Z"
-        clip-rule="evenodd"
-      />
-      <rect
-        width="184"
-        height="30"
-        x="786.695"
-        y="515.072"
-        fill="#929899"
-        rx="5"
-        transform="rotate(3 786.695 515.072)"
-      />
-      <rect
-        width="183"
-        height="29"
-        x="787.168"
-        y="515.597"
-        stroke="#68888F"
-        stroke-opacity=".8"
-        rx="4.5"
-        transform="rotate(3 787.168 515.597)"
-      />
-      <circle
-        cx="802.887"
-        cy="530.941"
-        r="10"
-        fill="#4ECB71"
-        transform="rotate(3 802.887 530.941)"
-      />
-      <path
-        fill="#fff"
-        fill-rule="evenodd"
-        d="M809.134 527.904c.085.077.153.169.201.27.048.101.074.21.077.321.004.11-.016.22-.057.323-.042.103-.104.197-.184.276l-7.475 7.445c-.084.083-.185.149-.297.194-.113.044-.234.065-.356.063-.122-.002-.243-.029-.355-.077-.111-.049-.211-.118-.293-.205l-3.94-4.149c-.154-.167-.233-.384-.22-.604.014-.22.117-.426.289-.572.172-.147.399-.223.631-.211.233.011.452.108.612.27l3.296 3.469 6.818-6.79c.161-.161.384-.253.619-.257.235-.004.463.079.634.234Z"
-        clip-rule="evenodd"
-      />
-      <rect
-        width="130"
-        height="13"
-        x="825.223"
-        y="525.102"
-        fill="#4ECB71"
-        rx="5"
-        transform="rotate(3 825.223 525.102)"
-      />
-      <defs>
-        <linearGradient
-          id="b"
-          x1="61.7591"
-          x2="669.906"
-          y1="18"
-          y2="389.116"
-          gradientUnits="userSpaceOnUse"
+      <Container
+        sx={{ display: 'flex', flexDirection: 'column', mt: 2, mb: 8 }}
+      >
+        <Typography
+          variant="h2"
+          sx={{
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'center',
+            alignItems: 'center',
+          }}
+          gutterBottom
         >
-          <stop stop-color="#176B87" stop-opacity=".5" />
-          <stop offset="1" stop-color="#176B87" stop-opacity=".2" />
-        </linearGradient>
-        <linearGradient
-          id="c"
-          x1="61.7591"
-          x2="648.581"
-          y1="27.728"
-          y2="369.89"
-          gradientUnits="userSpaceOnUse"
+          Master Competitive Programming
+        </Typography>
+        <Typography
+          variant="h4"
+          sx={{
+            fontSize: '2.5rem',
+            fontWeight: '400',
+            display: 'flex',
+            flexDirection: 'column',
+            justifyContent: 'center',
+            alignItems: 'center',
+          }}
         >
-          <stop stop-color="#0071A5" />
-          <stop offset="1" stop-color="#0071A5" />
-        </linearGradient>
-        <linearGradient
-          id="e"
-          x1="467.759"
-          x2="1075.91"
-          y1="238"
-          y2="609.116"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#176B87" stop-opacity=".7" />
-          <stop offset="1" stop-color="#176B87" stop-opacity=".5" />
-        </linearGradient>
-        <linearGradient
-          id="f"
-          x1="467.759"
-          x2="1054.58"
-          y1="247.728"
-          y2="589.89"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stop-color="#0071A5" />
-          <stop offset="1" stop-color="#0071A5" />
-        </linearGradient>
-        <filter
-          id="a"
-          width="648.599"
-          height="558.043"
-          x=".0376"
-          y=".03763"
-          color-interpolation-filters="sRGB"
-          filterUnits="userSpaceOnUse"
-        >
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feGaussianBlur in="BackgroundImageFix" stdDeviation="10" />
-          <feComposite
-            in2="SourceAlpha"
-            operator="in"
-            result="effect1_backgroundBlur_438_6"
+          Learn & Solve with O(fast)
+        </Typography>
+        <Box sx={{ height: '90vh' }}>
+          <img
+            className="card-rotate card-left"
+            src="assets/learn.svg"
+            alt="Learn card"
           />
-          <feColorMatrix
-            in="SourceAlpha"
-            result="hardAlpha"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
+          <img
+            className="card-rotate card-right"
+            src="assets/solve.svg"
+            alt="Solve card"
           />
-          <feOffset dy="20" />
-          <feGaussianBlur stdDeviation="20" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.2 0" />
-          <feBlend
-            in2="effect1_backgroundBlur_438_6"
-            result="effect2_dropShadow_438_6"
-          />
-          <feBlend
-            in="SourceGraphic"
-            in2="effect2_dropShadow_438_6"
-            result="shape"
-          />
-        </filter>
-        <filter
-          id="d"
-          width="648.599"
-          height="558.043"
-          x="406.038"
-          y="220.038"
-          color-interpolation-filters="sRGB"
-          filterUnits="userSpaceOnUse"
-        >
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feGaussianBlur in="BackgroundImageFix" stdDeviation="10" />
-          <feComposite
-            in2="SourceAlpha"
-            operator="in"
-            result="effect1_backgroundBlur_438_6"
-          />
-          <feColorMatrix
-            in="SourceAlpha"
-            result="hardAlpha"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-          />
-          <feOffset dy="20" />
-          <feGaussianBlur stdDeviation="20" />
-          <feComposite in2="hardAlpha" operator="out" />
-          <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.2 0" />
-          <feBlend
-            in2="effect1_backgroundBlur_438_6"
-            result="effect2_dropShadow_438_6"
-          />
-          <feBlend
-            in="SourceGraphic"
-            in2="effect2_dropShadow_438_6"
-            result="shape"
-          />
-        </filter>
-      </defs>
-    </svg>
+        </Box>
+      </Container>
+    </div>
   )
 }
